+++ pkg/limayaml/limayaml_test.go
@@ -21,7 +21,7 @@ const emptyYAML = "images: []\n"
 func TestEmptyYAML(t *testing.T) {
 	var y LimaYAML
 	t.Log(dumpJSON(t, y))
-	b, err := marshalYAML(y)
+	b, err := Marshal(&y, false)
 	assert.NilError(t, err)
 	assert.Equal(t, string(b), emptyYAML)
 }
@@ -32,12 +32,12 @@ func TestDefaultYAML(t *testing.T) {
 	bytes, err := os.ReadFile("default.yaml")
 	assert.NilError(t, err)
 	var y LimaYAML
-	err = unmarshalYAML(bytes, &y, "")
+	err = Unmarshal(bytes, &y, "")
 	assert.NilError(t, err)
 	y.Images = nil // remove default images
 	y.Mounts = nil // remove default mounts
 	t.Log(dumpJSON(t, y))
-	b, err := marshalYAML(y)
+	b, err := Marshal(&y, false)
 	assert.NilError(t, err)
 	assert.Equal(t, string(b), defaultYAML)
 }
