+++ pkg/limayaml/load.go
@@ -6,38 +6,10 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/goccy/go-yaml"
 	"github.com/lima-vm/lima/pkg/store/dirnames"
 	"github.com/lima-vm/lima/pkg/store/filenames"
-	"github.com/lima-vm/lima/pkg/yqutil"
 	"github.com/sirupsen/logrus"
 )
-
-func unmarshalDisk(dst *Disk, b []byte) error {
-	var s string
-	if err := yaml.Unmarshal(b, &s); err == nil {
-		*dst = Disk{Name: s}
-		return nil
-	}
-	return yaml.Unmarshal(b, dst)
-}
-
-func unmarshalYAML(data []byte, v interface{}, comment string) error {
-	if err := yaml.UnmarshalWithOptions(data, v, yaml.DisallowDuplicateKey(), yaml.CustomUnmarshaler[Disk](unmarshalDisk)); err != nil {
-		return fmt.Errorf("failed to unmarshal YAML (%s): %w", comment, err)
-	}
-	// the go-yaml library doesn't catch all markup errors, unfortunately
-	// make sure to get a "second opinion", using the same library as "yq"
-	if err := yqutil.ValidateContent(data); err != nil {
-		return fmt.Errorf("failed to unmarshal YAML (%s): %w", comment, err)
-	}
-	if err := yaml.UnmarshalWithOptions(data, v, yaml.Strict(), yaml.CustomUnmarshaler[Disk](unmarshalDisk)); err != nil {
-		logrus.WithField("comment", comment).WithError(err).Warn("Non-strict YAML is deprecated and will be unsupported in a future version of Lima")
-		// Non-strict YAML is known to be used by Rancher Desktop:
-		// https://github.com/rancher-sandbox/rancher-desktop/blob/c7ea7508a0191634adf16f4675f64c73198e8d37/src/backend/lima.ts#L114-L117
-	}
-	return nil
-}
 
 // Load loads the yaml and fulfills unspecified fields with the default values.
 //
@@ -45,7 +17,7 @@ func unmarshalYAML(data []byte, v interface{}, comment string) error {
 func Load(b []byte, filePath string) (*LimaYAML, error) {
 	var y, d, o LimaYAML
 
-	if err := unmarshalYAML(b, &y, fmt.Sprintf("main file %q", filePath)); err != nil {
+	if err := Unmarshal(b, &y, fmt.Sprintf("main file %q", filePath)); err != nil {
 		return nil, err
 	}
 	configDir, err := dirnames.LimaConfigDir()
@@ -57,7 +29,7 @@ func Load(b []byte, filePath string) (*LimaYAML, error) {
 	bytes, err := os.ReadFile(defaultPath)
 	if err == nil {
 		logrus.Debugf("Mixing %q into %q", defaultPath, filePath)
-		if err := unmarshalYAML(bytes, &d, fmt.Sprintf("default file %q", defaultPath)); err != nil {
+		if err := Unmarshal(bytes, &d, fmt.Sprintf("default file %q", defaultPath)); err != nil {
 			return nil, err
 		}
 	} else if !errors.Is(err, os.ErrNotExist) {
@@ -68,7 +40,7 @@ func Load(b []byte, filePath string) (*LimaYAML, error) {
 	bytes, err = os.ReadFile(overridePath)
 	if err == nil {
 		logrus.Debugf("Mixing %q into %q", overridePath, filePath)
-		if err := unmarshalYAML(bytes, &o, fmt.Sprintf("override file %q", overridePath)); err != nil {
+		if err := Unmarshal(bytes, &o, fmt.Sprintf("override file %q", overridePath)); err != nil {
 			return nil, err
 		}
 	} else if !errors.Is(err, os.ErrNotExist) {
